Add day 18
Complete rewrite from part 1: using sets is impossible for part 2, so use a formula from https://mathopenref.com/coordpolygonarea.html
Use complex numbers instead of tuples, because adding and multiplying is so much easier this way.

The discreteness of the problem means some tweaks are needed to count the boundary properly: if you image the vertices being on the top left corner, then the top and left boundaries are counted, but not the bottom and right. The final +1 is probably the bottom right corner in this scenario, but I'm not sure.